Move Bike Views 'book' and 'plot' to 'bikes.views'
